Cicd-177: Add static files to ECS site CDN

+++ src/aws/cloudformation/delete-ecs-cloudfront-distribution-stack.sh
@@ -8,6 +8,15 @@ cd $(dirname "$0")
 source ../aws-functions.sh
 source ../../compute-variables.sh
 
+# If the bucket exists, empty it; otherwise, CloudFormation won't be able to delete it
+../s3/empty-bucket.sh ${ProjectBucketName} 'site'
+if [[ $? -ne 0 ]]
+then
+  echo 'Deletion of the stack has been aborted.'
+  exit 1
+fi
+
+echo 'Requesting deletion of the stack...'
 OUTPUT=$(aws cloudformation delete-stack \
   --profile ${Profile} \
   --region ${Region} \
